+++ Runtime/Scripts/GlobalTools/UI/UIFPS.cs
@@ -3,14 +3,16 @@ using System.Collections.Generic;
 using UnityEngine;
 
 using TMPro;
+using static PolytopeSolutions.Toolset.GlobalTools.Generic.ObjectHelpers;
 
 namespace PolytopeSolutions.Toolset.GlobalTools.UI {
 	[ExecuteAlways]
-	[RequireComponent(typeof(TextMeshProUGUI))]
 	public class UIFPS : MonoBehaviour {
-		private TextMeshProUGUI label;
-		[Range(0,6)]
+        [SerializeField] private TextMeshProUGUI averageFPSLabel;
+        [SerializeField] private TextMeshProUGUI lowestFPSLabel;
+        [Range(0,6)]
 		[SerializeField] private int precision = 3;
+		[SerializeField] private bool unscaledDeltaTime = false;
 
 		private float[] buffer;
 		[SerializeField] private int bufferSize = 10;
@@ -25,6 +27,15 @@ namespace PolytopeSolutions.Toolset.GlobalTools.UI {
 				return sum/count;
 			}
 		}
+		private float lowestFPS {
+            get {
+                float lowest = float.MaxValue;
+                int count = (this.filled) ? this.bufferSize : this.currentIndex;
+                for (int i = 0; i < count; i++)
+                    lowest = Mathf.Min(lowest, this.buffer[i]);
+                return lowest;
+            }
+        }
 
 		private void OnEnable() {
 			Initizalize();
@@ -33,7 +44,8 @@ namespace PolytopeSolutions.Toolset.GlobalTools.UI {
 			Initizalize();
 		}
         private void Initizalize() {
-			this.label = GetComponent<TextMeshProUGUI>();
+			if (!this.averageFPSLabel)
+				this.averageFPSLabel = gameObject.TryGetOrAddComponent<TextMeshProUGUI>();
 			this.buffer = new float[this.bufferSize];
 			this.currentIndex = 0;
 			this.filled = false;
@@ -41,13 +53,15 @@ namespace PolytopeSolutions.Toolset.GlobalTools.UI {
 
 		private void Update() {
 			if (this.buffer != null) {
-				float value = 1.0f / Time.deltaTime;
+				float value = 1.0f / ((this.unscaledDeltaTime) ? Time.unscaledDeltaTime : Time.deltaTime);
 				this.buffer[this.currentIndex] = value;
 				this.currentIndex = (this.currentIndex + 1) % this.bufferSize;
 				if (this.currentIndex == 0) this.filled = true;
 
-				this.label.text = this.averageFPS.ToString("F" + this.precision);
-			}
+				this.averageFPSLabel.text = this.averageFPS.ToString($"F{this.precision}");
+                if (this.lowestFPSLabel)
+					this.lowestFPSLabel.text = this.lowestFPS.ToString($"F{this.precision}");
+            }
 		}
 	}
 }